+++ backend/src/main/java/com/jwt/spring_security/repo/EmployeeRepo.java
@@ -4,12 +4,10 @@ import com.jwt.spring_security.model.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
-
 @Repository
 public interface EmployeeRepo extends JpaRepository<Employee, Long> {
 
-    Employee findByEmployeeID(int employeeID);
+    Employee findByEmployeeID(Long employeeID);
 
 
 }
